requirements.txt
numpy>=1.24
shapely>=2.0
rasterio>=1.3
opencv-python>=4.8
Pillow>=10.0
pydantic>=2.0
